Add support for multiple methods from client
As per RFC1928, we need to be able to receive multiple methods reported
by the client. As such, we need to read nmethods bits from the socket
for methods, even if we ignore them and only support no authentication.
Otherwise, this was causing errors for the next reads when the client
sent more than one supported method.